rename status -> dealStatus
