+++ Lighthouse/dealStatus/index.js
@@ -3,13 +3,13 @@ const lighthouseConfig = require("../../lighthouse.config");
 
 module.exports = async (cid) => {
   try {
-    const status = (
+    const dealStatus = (
       await axios.get(
         lighthouseConfig.lighthouseAPI +
-          `/api/lighthouse/cid_status/?cid=${cid}`
+          `/api/lighthouse/deal_status/?cid=${cid}`
       )
     ).data;
-    return {data: { status }};
+    return {data: { dealStatus }};
   } catch (error) {
     throw new Error(error.message);
   }
